student base funtion has been done, fast and fast finish project

+++ campus/physicsExperiment/src/main/java/edu/cumt/phyExperiment/service/CollegeService.java
@@ -13,9 +13,23 @@ import java.util.List;
  */
 public interface CollegeService {
 
-    String queryCollegeNameById(int collegeId);
+    /**
+     * 根据学院代号查询学院名称
+     * @param collegeId 学院代号
+     * @return 返回学院名称或null
+     */
+    String queryCollegeNameById(long collegeId);
 
+    /**
+     * 查询所有学院信息
+     * @return
+     */
     List<College> queryAllCollegesInfo();
 
-    String queryCollegeDescriptionById(int collegeId);
+    /**
+     * 查询学院的介绍
+     * @param collegeId
+     * @return
+     */
+    String queryCollegeDescriptionById(long collegeId);
 }
